Core/Player: Fixed currentPetSlot loading and saving.

+++ sql/base/characters_database.sql
@@ -838,7 +838,7 @@ CREATE TABLE `characters` (
   `ammoId` int(10) unsigned NOT NULL DEFAULT '0',
   `knownTitles` longtext,
   `actionBars` tinyint(3) unsigned NOT NULL DEFAULT '0',
-  `currentPetSlot` int(10) DEFAULT NULL,
+  `currentPetSlot` tinyint(3) DEFAULT NULL,
   `petSlotUsed` int(10) DEFAULT NULL,
   `grantableLevels` tinyint(3) unsigned NOT NULL DEFAULT '0',
   `deleteInfos_Account` int(10) unsigned DEFAULT NULL,
